requirements.txt
APScheduler==3.6.3
black==22.6.0
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
flake8==5.0.4
flake8-broken-line==0.5.0
flake8-isort==4.2.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
idna==3.3
isort==5.10.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
pathspec==0.9.0
pep8-naming==0.13.2
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tomli==2.0.1
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
